requirements.txt
PyQt5>=5.15.0
skia-python>=87.0